refactor: move forest logic to starknet crate
