+++ crates/committer/src/lib.rs
@@ -1,6 +1,4 @@
-pub mod block_committer;
 pub mod felt;
-pub mod forest_errors;
 pub mod hash;
 pub mod patricia_merkle_tree;
 pub mod storage;
