+++ crates/committer/src/patricia_merkle_tree/filled_tree.rs
@@ -1,5 +1,4 @@
 pub mod errors;
-pub mod forest;
 pub mod node;
 pub mod node_serde;
 pub mod tree;
