+++ crates/committer/src/patricia_merkle_tree/original_skeleton_tree.rs
@@ -2,6 +2,5 @@ pub mod config;
 pub mod create_tree;
 pub mod errors;
 pub mod node;
-pub mod skeleton_forest;
 pub mod tree;
 pub mod utils;
