+++ crates/committer/src/patricia_merkle_tree/updated_skeleton_tree.rs
@@ -2,5 +2,4 @@ pub mod create_tree_helper;
 pub mod errors;
 pub mod hash_function;
 pub mod node;
-pub mod skeleton_forest;
 pub mod tree;
