+++ crates/starknet-committer/src/lib.rs
@@ -1,2 +1,3 @@
 pub mod block_committer;
+pub mod starknet_forest;
 pub mod starknet_patricia_merkle_tree;
