+++ crates/starknet-committer/src/starknet_forest.rs
@@ -1,0 +1,4 @@
+pub mod filled_forest;
+pub mod forest_errors;
+pub mod original_skeleton_forest;
+pub mod updated_skeleton_forest;
