+++ crates/starknet-committer/src/starknet_forest/filled_forest.rs
@@ -1,25 +1,26 @@
 use std::collections::HashMap;
 use std::sync::Arc;
 
+use committer::hash::hash_trait::HashOutput;
+use committer::patricia_merkle_tree::filled_tree::tree::FilledTree;
+use committer::patricia_merkle_tree::node_data::leaf::LeafModifications;
+use committer::patricia_merkle_tree::types::NodeIndex;
+use committer::patricia_merkle_tree::updated_skeleton_tree::tree::UpdatedSkeletonTreeImpl;
+use committer::storage::storage_trait::Storage;
 use tokio::task::JoinSet;
 
 use crate::block_committer::input::{ContractAddress, StarknetStorageValue};
-use crate::forest_errors::{ForestError, ForestResult};
-use crate::hash::hash_trait::HashOutput;
-use crate::patricia_merkle_tree::filled_tree::node::{ClassHash, CompiledClassHash, Nonce};
-use crate::patricia_merkle_tree::filled_tree::tree::{
+use crate::hash::ForestHashFunction;
+use crate::starknet_forest::forest_errors::{ForestError, ForestResult};
+use crate::starknet_forest::updated_skeleton_forest::UpdatedSkeletonForest;
+use crate::starknet_patricia_merkle_tree::node::{ClassHash, CompiledClassHash, Nonce};
+use crate::starknet_patricia_merkle_tree::starknet_leaf::leaf::ContractState;
+use crate::starknet_patricia_merkle_tree::types::{
     ClassesTrie,
     ContractsTrie,
-    FilledTree,
     StorageTrie,
     StorageTrieMap,
 };
-use crate::patricia_merkle_tree::node_data::leaf::{ContractState, LeafModifications};
-use crate::patricia_merkle_tree::types::NodeIndex;
-use crate::patricia_merkle_tree::updated_skeleton_tree::hash_function::ForestHashFunction;
-use crate::patricia_merkle_tree::updated_skeleton_tree::skeleton_forest::UpdatedSkeletonForest;
-use crate::patricia_merkle_tree::updated_skeleton_tree::tree::UpdatedSkeletonTreeImpl;
-use crate::storage::storage_trait::Storage;
 
 pub struct FilledForest {
     pub storage_tries: StorageTrieMap,
